# apps/App/src/middleware/tsoa/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"PatientCheckup": {
				"properties": {
					"id": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"bloodPressure": {
						"type": "string"
					},
					"cholestrol": {
						"type": "string"
					},
					"glucose": {
						"type": "string"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"weight": {
						"type": "number",
						"format": "double"
					},
					"doctor": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"type",
					"bloodPressure",
					"cholestrol",
					"glucose",
					"height",
					"weight",
					"doctor"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PatientBase": {
				"properties": {
					"id": {
						"type": "string"
					},
					"type": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PatientInfo": {
				"properties": {
					"id": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"bloodType": {
						"type": "string"
					},
					"gender": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					},
					"race": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"type",
					"firstName",
					"lastName",
					"bloodType",
					"gender",
					"age",
					"race"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"in": "header",
				"name": "token"
			}
		}
	},
	"info": {
		"title": "Cross_Region_Active-Active",
		"version": "0.0.1",
		"description": "Cross Region Active-Active Backend (App)",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Martin Greyling"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/patient/initialize": {
			"post": {
				"operationId": "Initialize",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"summary": "Get a Consumer's daily quotas",
				"tags": [
					"Patient"
				],
				"security": [],
				"parameters": []
			}
		},
		"/patient": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Patient"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "Post",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Patient"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"anyOf": [
									{
										"$ref": "#/components/schemas/PatientCheckup"
									},
									{
										"$ref": "#/components/schemas/PatientBase"
									},
									{
										"$ref": "#/components/schemas/PatientInfo"
									}
								]
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}